+++ pkg/appsec/mocks.go
@@ -288,6 +288,14 @@ func (m *Mock) UpdateCustomRule(ctx context.Context, req UpdateCustomRuleRequest
 	return args.Get(0).(*UpdateCustomRuleResponse), args.Error(1)
 }
 
+func (m *Mock) GetCustomRulesUsage(ctx context.Context, req GetCustomRulesUsageRequest) (*GetCustomRulesUsageResponse, error) {
+	args := m.Called(ctx, req)
+	if args.Get(0) == nil {
+		return nil, args.Error(1)
+	}
+	return args.Get(0).(*GetCustomRulesUsageResponse), args.Error(1)
+}
+
 func (m *Mock) UpdateCustomDeny(ctx context.Context, req UpdateCustomDenyRequest) (*UpdateCustomDenyResponse, error) {
 	args := m.Called(ctx, req)
 	if args.Get(0) == nil {